changed all styles to styles.jsx , and fix other comments
