+++ src/index.scss
@@ -1,13 +1,11 @@
 body {
   margin: 0;
-  font-family: 'Montserrat', sans-serif;
+  font-family: 'Aribau Grotesk,Montserrat', sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
 }
 
 code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
+  font-family: Aribau Grotesk, source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
     monospace;
 }
-
-
